pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.57.1",
    "discord-py>=2.5.2",
    "google-genai>=1.24.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]
